+++ bindings/matrix-sdk-ffi/src/api.udl
@@ -1,19 +1,5 @@
 namespace matrix_sdk_ffi {};
 
-enum SlidingSyncListLoadingState {
-    /// Sliding Sync has not started to load anything yet.
-    "NotLoaded",
-    /// Sliding Sync has been preloaded, i.e. restored from a cache for example.
-    "Preloaded",
-    /// We are trying to load all remaining rooms, might be in batches
-    /// Updates are received from the loaded rooms, and new rooms are being fetched
-    /// in background
-    "PartiallyLoaded",
-    /// Updates are received for all the loaded rooms, and all rooms have been
-    /// loaded!
-    "FullyLoaded",
-};
-
 dictionary CreateRoomParameters {
     string? name;
     string? topic = null;
